RTE-1001: Initial tdx converter implementation.
